Improve tracing setup
