+++ apps/hash-external-services/docker-compose.dev.yml
@@ -88,10 +88,8 @@ services:
     depends_on:
       telemetry-collector:
         condition: service_healthy
-    # When an error happens on the graph, OTLP will run into a data race
-    # and crash the graph. For now, we've disabled traces.
-    # environment:
-    #   HASH_GRAPH_OTLP_ENDPOINT: "http://telemetry-collector:4317"
+    environment:
+      HASH_GRAPH_OTLP_ENDPOINT: "http://telemetry-collector:4317"
 
   graph:
     depends_on:
@@ -109,10 +107,7 @@ services:
       HASH_SPICEDB_GRPC_PRESHARED_KEY: "${HASH_SPICEDB_GRPC_PRESHARED_KEY}"
       HASH_TEMPORAL_SERVER_HOST: "http://temporal"
       HASH_TEMPORAL_SERVER_PORT: "${HASH_TEMPORAL_SERVER_PORT}"
-    # When an error happens on the graph, OTLP will run into a data race
-    # and crash the graph. For now, we've disabled traces.
-    # environment:
-    #   HASH_GRAPH_OTLP_ENDPOINT: "http://telemetry-collector:4317"
+      HASH_GRAPH_OTLP_ENDPOINT: "http://telemetry-collector:4317"
     ports:
       - "${HASH_GRAPH_API_PORT}:4000"
