+++ apps/hash-graph/bins/cli/Cargo.toml
@@ -16,6 +16,7 @@ type-fetcher = { workspace = true }
 temporal-client = { workspace = true }
 authorization = { workspace = true }
 codec = { workspace = true }
+hash-tracing = { workspace = true, features = ["clap"] }
 
 error-stack = { workspace = true }
 type-system = { workspace = true }
@@ -28,7 +29,6 @@ futures = { version = "0.3.30" }
 regex = "1.10.2"
 reqwest = { version = "0.11", default-features = false, features = ["rustls-tls"] }
 semver = { version = "1.0.21", default-features = false }
-sentry = { version = "0.32.1", features = ["backtrace", "contexts", "debug-images", "panic", "reqwest", "rustls"], default-features = false }
 serde_json = "1.0.111"
 tarpc = { version = "0.33", features = ["serde1", "tokio1", "serde-transport", "tcp"] }
 time = "0.3.31"
@@ -36,7 +36,7 @@ tokio = { workspace = true }
 tokio-postgres = { version = "0.7.10", default-features = false }
 tokio-serde = { version = "0.8", features = ["json"] }
 tokio-util = { workspace = true, features = ["codec"] }
-tracing = "0.1.40"
+tracing = { workspace = true }
 uuid = "1.6.1"
 
 [features]
