+++ apps/hash-graph/bins/cli/src/main.rs
@@ -8,13 +8,11 @@
 
 mod args;
 mod error;
-mod parser;
 mod subcommand;
-
-use std::{borrow::Cow, sync::Arc};
 
 use error_stack::Result;
 use graph::load_env;
+use hash_tracing::sentry::init_sentry;
 
 use self::{args::Args, error::GraphError};
 
@@ -26,31 +24,11 @@ fn main() -> Result<(), GraphError> {
     validation::error::install_error_stack_hooks();
 
     let Args {
-        sentry_dsn,
-        sentry_environment,
         subcommand,
+        tracing_config,
     } = Args::parse_args();
 
-    // Initialize Sentry
-    // When initializing Sentry, a `Drop` guard is returned, once dropped any remaining events are
-    // flushed. This means we need to keep the guard around for the entire lifetime of the program.
-    let _sentry = sentry::init(sentry::ClientOptions {
-        dsn: sentry_dsn,
-        release: sentry::release_name!(),
-        session_mode: sentry::SessionMode::Request,
-        traces_sampler: Some(Arc::new(|ctx| {
-            if Some(true) == ctx.sampled() {
-                1.0
-            } else if ctx.operation() == "http.server" {
-                0.1
-            } else {
-                1.0
-            }
-        })),
-        environment: Some(Cow::Owned(sentry_environment.to_string())),
+    let _sentry_guard = init_sentry(&tracing_config.sentry);
 
-        ..Default::default()
-    });
-
-    subcommand.execute()
+    subcommand.execute(tracing_config)
 }
