+++ apps/hash-graph/bins/cli/src/subcommand/mod.rs
@@ -9,6 +9,7 @@ mod type_fetcher;
 use std::future::Future;
 
 use error_stack::Result;
+use hash_tracing::{init_tracing, TracingConfig};
 
 #[cfg(feature = "test-server")]
 pub use self::test_server::{test_server, TestServerArgs};
@@ -39,27 +40,33 @@ pub enum Subcommand {
     TestServer(TestServerArgs),
 }
 
-fn block_on(future: impl Future<Output = Result<(), GraphError>>) -> Result<(), GraphError> {
+fn block_on(
+    future: impl Future<Output = Result<(), GraphError>>,
+    tracing_config: TracingConfig,
+) -> Result<(), GraphError> {
     tokio::runtime::Builder::new_multi_thread()
         .enable_all()
         .build()
         .expect("failed to create runtime")
-        .block_on(future)
+        .block_on(async {
+            let _log_guard = init_tracing(tracing_config);
+            future.await
+        })
 }
 
 impl Subcommand {
-    pub(crate) fn execute(self) -> Result<(), GraphError> {
+    pub(crate) fn execute(self, tracing_config: TracingConfig) -> Result<(), GraphError> {
         match self {
-            Self::Server(args) => block_on(server(args)),
-            Self::Migrate(args) => block_on(migrate(args)),
-            Self::TypeFetcher(args) => block_on(type_fetcher(args)),
+            Self::Server(args) => block_on(server(args), tracing_config),
+            Self::Migrate(args) => block_on(migrate(args), tracing_config),
+            Self::TypeFetcher(args) => block_on(type_fetcher(args), tracing_config),
             Self::Completions(ref args) => {
                 completions(args);
                 Ok(())
             }
-            Self::Snapshot(args) => block_on(snapshot(args)),
+            Self::Snapshot(args) => block_on(snapshot(args), tracing_config),
             #[cfg(feature = "test-server")]
-            Self::TestServer(args) => block_on(test_server(args)),
+            Self::TestServer(args) => block_on(test_server(args), tracing_config),
         }
     }
 }
