+++ apps/hash-graph/libs/api/Cargo.toml
@@ -38,7 +38,7 @@ tokio-postgres = { version = "0.7.10", default-features = false, optional = true
 tokio-util = { workspace = true, features = ["io"], optional = true }
 tower = "0.4.13"
 tower-http = { version = "0.5.1", features = ["trace"] }
-tracing = "0.1.40"
+tracing = { workspace = true }
 tracing-opentelemetry = "0.22.0"
 utoipa = "4.2.0"
 uuid = "1.6.1"
