--- apps/hash-graph/libs/graph/src/logging/mod.rs
@@ -1,5 +1,0 @@
-mod args;
-mod init;
-
-pub use args::LoggingArgs;
-pub use init::init_logger;
