+++ apps/hash-graph/libs/type-fetcher/Cargo.toml
@@ -10,7 +10,7 @@ type-system = { workspace = true }
 
 serde = { version = "1.0.195", features = ["derive"] }
 time = { workspace = true, features = ["serde", "std"] }
-tracing = "0.1.40"
+tracing = { workspace = true }
 
 futures = "0.3"
 tarpc = { version = "0.33", features = ["tokio1"] }
