+++ apps/hash-graph/package.json
@@ -34,6 +34,7 @@
     "@local/status-rust": "0.0.0-private",
     "@local/temporal-client-rs": "0.0.0-private",
     "@local/temporal-versioning-rs": "0.0.0-private",
+    "@local/tracing-rs": "0.0.0-private",
     "ts-node": "10.9.1"
   },
   "devDependencies": {
