+++ libs/@local/temporal-versioning/Cargo.toml
@@ -12,7 +12,7 @@ bytes = { version = "1.5.0", optional = true }
 utoipa = { version = "4.2.0", optional = true }
 postgres-types = { version = "0.2.6", default-features = false, features = ["with-time-0_3"], optional = true }
 time = { version = "0.3.31", default-features = false, features = ["serde", "parsing", "formatting", "macros"] }
-tracing = { version = "0.1.40", optional = true }
+tracing = { workspace = true, optional = true }
 
 derivative = "2.2.0"
 postgres-protocol = { version = "0.6.6", default-features = false, optional = true }
