+++ libs/@local/tracing/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@local/tracing-rs",
+  "version": "0.0.0-private",
+  "private": true,
+  "scripts": {
+    "fix:clippy": "just clippy --fix",
+    "lint:clippy": "just clippy"
+  }
+}
