+++ libs/error-stack/Cargo.toml
@@ -29,7 +29,7 @@ serde = { version = "1.0.195", features = ["derive"] }
 serde_json = "1.0.111"
 futures = { version = "0.3.30", default-features = false, features = ["executor"] }
 trybuild = "1.0.89"
-tracing = "0.1.40"
+tracing = { workspace = true }
 tracing-subscriber = "0.3.18"
 insta = { version = "1.34.0", features = ['filters', 'ron'] }
 regex = "1.10.2"
